refactor: move app/auth folder under app/pages
