+++ src/app/pages/auth/auth-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { AuthPageComponent } from './pages/auth-page/auth-page.component';
-import { LoginFormComponent } from './forms/login-form/login-form.component';
-import { SignupFormComponent } from './forms/signup-form/signup-form.component';
+import { AuthPageComponent } from './components/auth-page/auth-page.component';
+import { LoginFormComponent } from './components/login-form/login-form.component';
+import { SignupFormComponent } from './components/signup-form/signup-form.component';
 
 const routes: Routes = [
   {
