+++ src/app/pages/auth/components/auth-page/auth-page.component.scss
