+++ src/app/pages/auth/components/signup-form/signup-form.component.scss
